add user edit functionality

+++ frontend/src/screens/UserListScreen.js
@@ -34,7 +34,9 @@ function UserListScreen() {
     },[dispatch, navigate,successDelete])
 
     const deleteHandler = (id) => {
-      dispatch(deleteUser(id))
+      if(window.confirm('Are you sure you want to delete this user? ')){
+        dispatch(deleteUser(id))
+      }
     }
 
   return (
@@ -68,7 +70,7 @@ function UserListScreen() {
                       <i className='fas fa-x' style={{color: 'red'}}></i>
                     )}</td>
                     <td>
-                      <LinkContainer to={`/admin/user/${user._id}`}>
+                      <LinkContainer to={`/admin/user/${user._id}/edit`}>
                         <Button variant='light' className='btn-sm'>
                           <i className="fas fa-edit"></i>
                         </Button>
